feat: implement platform-specific entry points for web and Tauri builds
- Split setup.js into platform-specific files (web/tauri/common)
- Add build script to switch between platforms at build time
- Update package.json scripts for separate web and Tauri builds
- Add setup.js to .gitignore as it's now a generated file
- Update README with Tauri desktop application instructions

This allows the app to have different configurations and behaviors
when running as a web app vs desktop app via Tauri.
